Moved hr_slice into mcmc
